Combines all the document apis into one object. Splits the test.
