--- tests/unit/api/test_documents.py
@@ -1,95 +1,0 @@
-import unittest
-
-from http import HTTPStatus
-from mock import patch, Mock
-
-from search_service.api.documents import DocumentTablesAPI, DocumentUsersAPI
-from search_service import create_app
-
-
-class TestDocumentTablesAPI(unittest.TestCase):
-    def setUp(self) -> None:
-        self.app = create_app(config_module_class='search_service.config.Config')
-        self.app_context = self.app.app_context()
-        self.app_context.push()
-
-    def tear_down(self):
-        self.app_context.pop()
-
-    @patch('search_service.api.documents.TableSchema')
-    @patch('search_service.api.documents.reqparse.RequestParser')
-    @patch('search_service.api.documents.get_proxy_client')
-    def test_post(self, get_proxy, RequestParser, TableSchema) -> None:
-        mock_proxy = get_proxy.return_value = Mock()
-        RequestParser().parse_args.return_value = dict(data='{}', index='fake_index')
-        expected_value = TableSchema().loads.return_value = Mock()
-
-        response = DocumentTablesAPI().post()
-        self.assertEqual(list(response)[1], HTTPStatus.OK)
-        mock_proxy.create_document.assert_called_with(data=expected_value, index='fake_index')
-
-    @patch('search_service.api.documents.TableSchema')
-    @patch('search_service.api.documents.reqparse.RequestParser')
-    @patch('search_service.api.documents.get_proxy_client')
-    def test_put(self, get_proxy, RequestParser, TableSchema) -> None:
-        mock_proxy = get_proxy.return_value = Mock()
-        RequestParser().parse_args.return_value = dict(data='{}', index='fake_index')
-        expected_value = TableSchema().loads.return_value = Mock()
-
-        response = DocumentTablesAPI().put()
-        self.assertEqual(list(response)[1], HTTPStatus.OK)
-        mock_proxy.update_document.assert_called_with(data=expected_value, index='fake_index')
-
-    def test_should_not_reach_create_with_id(self):
-        response = self.app.test_client().post('/document_table/1')
-
-        self.assertEquals(response.status_code, 405)
-
-    def test_should_not_reach_update_with_id(self):
-        response = self.app.test_client().put('/document_table/1')
-
-        self.assertEquals(response.status_code, 405)
-
-
-class DocumentUsersAPITest(unittest.TestCase):
-    def setUp(self) -> None:
-        self.app = create_app(config_module_class='search_service.config.Config')
-        self.app_context = self.app.app_context()
-        self.app_context.push()
-
-    def tear_down(self):
-        self.app_context.pop()
-
-    @patch('search_service.api.documents.UserSchema')
-    @patch('search_service.api.documents.reqparse.RequestParser')
-    @patch('search_service.api.documents.get_proxy_client')
-    def test_post(self, get_proxy, RequestParser, UserSchema) -> None:
-        mock_proxy = get_proxy.return_value = Mock()
-        RequestParser().parse_args.return_value = dict(data='{}', index='fake_index')
-        expected_value = UserSchema().loads.return_value = Mock()
-
-        response = DocumentUsersAPI().post()
-        self.assertEqual(list(response)[1], HTTPStatus.OK)
-        mock_proxy.create_document.assert_called_with(data=expected_value, index='fake_index')
-
-    @patch('search_service.api.documents.UserSchema')
-    @patch('search_service.api.documents.reqparse.RequestParser')
-    @patch('search_service.api.documents.get_proxy_client')
-    def test_put(self, get_proxy, RequestParser, UserSchema) -> None:
-        mock_proxy = get_proxy.return_value = Mock()
-        RequestParser().parse_args.return_value = dict(data='{}', index='fake_index')
-        expected_value = UserSchema().loads.return_value = Mock()
-
-        response = DocumentUsersAPI().put()
-        self.assertEqual(list(response)[1], HTTPStatus.OK)
-        mock_proxy.update_document.assert_called_with(data=expected_value, index='fake_index')
-
-    def test_should_not_reach_create_with_id(self):
-        response = self.app.test_client().post('/document_user/1')
-
-        self.assertEquals(response.status_code, 405)
-
-    def test_should_not_reach_update_with_id(self):
-        response = self.app.test_client().put('/document_user/1')
-
-        self.assertEquals(response.status_code, 405)
